backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "wildlife-camera-processor"
version = "0.1.0"
description = "A Python CLI tool for processing wildlife camera images using PyTorch Wildlife"
authors = [
    {name = "Wildlife Research Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytorchwildlife>=1.0.0",
    "pillow>=10.0.0",
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
    "lightning>=2.3.3",
    "omegaconf>=2.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "huggingface-hub>=1.1.2",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "flower>=2.0.0",
    "pyjwt[crypto]>=2.8.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-xdist>=3.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "httpx>=0.25.0",
    "pre-commit>=3.0.0",
    "ruff>=0.13.3",
    "types-python-dateutil>=2.9.0.20251115",
    "types-requests>=2.31.0.0",
    "types-beautifulsoup4>=4.12.0.0",
]

[project.scripts]
wildlife-processor = "wildlife_processor.cli.main:app"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "types-python-dateutil>=2.9.0.20251115",
    "types-requests>=2.31.0.0",
    "types-beautifulsoup4>=4.12.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
mypy_path = "src"
exclude = [
    "^build/",
    "^dist/",
    "^\\.venv/",
]

[[tool.mypy.overrides]]
module = [
    "torchvision.*",
    "PytorchWildlife.*",
    "bs4.*",
]
ignore_missing_imports = true
